* Bitmap font ImportFolder, missing trim and toalpha

+++ Source/Gorgon/Graphics/Color.h
@@ -301,7 +301,7 @@ namespace Gorgon { namespace Graphics {
 		explicit RGBAf(float lum, float a=1.0f) : RGBAf(lum, lum, lum, a) { }
 
 		/// Constructor that sets all color channels to the given value to create a grayscale color. Alpha is set to 1.0f
-		explicit RGBAf(double lum, float a=1.0f) : RGBAf(lum, lum, lum, a) { }
+		explicit RGBAf(double lum, float a=1.0f) : RGBAf((float)lum, (float)lum, (float)lum, a) { }
 
 		/// Converts a RGBA to RGBAf
 		RGBAf(const RGBA &color) : R(color.R/255.f), G(color.G/255.f), B(color.B/255.f), A(color.A/255.f) { }
